add json export, clean up code, fix auth

+++ app/Http/Controllers/GuestbookController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\guestbook;
+use App\Models\Guestbook;
 use Illuminate\Http\Request;
 
 class GuestbookController extends Controller
@@ -44,14 +44,18 @@ class GuestbookController extends Controller
      */
     public function show(guestbook $guestbook)
     {
-        return view('guestbooks.show');
+        abort(404);
     }
 
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(guestbook $guestbook)
+    public function edit(Request $request, guestbook $guestbook)
     {
+        if ($request->user()->cannot('update', $guestbook)) {
+            abort(403);
+        }
+
         return view('guestbooks.edit', compact(["guestbook"]));
     }
 
@@ -60,6 +64,10 @@ class GuestbookController extends Controller
      */
     public function update(Request $request, guestbook $guestbook)
     {
+        if ($request->user()->cannot('update', $guestbook)) {
+            abort(403);
+        }
+
         $validated = $request->validate([
             'name' =>'required|max:255',
             'style' => 'nullable'
@@ -73,13 +81,25 @@ class GuestbookController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(guestbook $guestbook)
+    public function destroy(Request $request, guestbook $guestbook)
     {
+        if ($request->user()->cannot('delete', $guestbook)) {
+            abort(403);
+        }
+
         $guestbook->delete();
         return redirect()->route('guestbooks.index')->with('success','Guestbook deleted successfully');
     }
 
-    public function delete(guestbook $guestbook) {
+    public function delete(Request $request, guestbook $guestbook) {
+        // Our delete route is custom and has a ->can(delete) instruction,
+        // so technically we don't need to do this...
+        // but I'd rather have it be obvious that we're doing authorization when looking at either routes/web or the actual controller
+        // Anything else would look inconsistent, and I don't want to accidentally end up removing both instructions.
+        if ($request->user()->cannot('delete', $guestbook)) {
+            abort(403);
+        }
+
         return view('guestbooks.delete', compact(['guestbook']));
     }
 }
